# plugins/rest-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zonvo AI REST API
  description: |
    Welcome to the Zonvo AI REST API. Everything your AI voice agents do on the phone — calling, qualifying,
    booking, promising a callback, honouring a do-not-call request — is available to your CRM, ERP or
    in-house app through `https://app.zonvo.tech/api/v1`.

    ## What you can automate

    | Goal | Endpoints |
    |------|-----------|
    | **Trigger calls** from your own app, now or at a scheduled time | `POST /calls` (with `scheduledAt`, `dynamicVariables`) |
    | **Run bulk campaigns** | `POST /campaigns`, `/campaigns/{id}/contacts`, `/start`, `/pause` |
    | **Sync leads and call outcomes to your CRM** | `GET /leads`, `PATCH /leads/{id}`, `POST /leads/{id}/notes`, `GET /calls?outcome=`, `POST /calls/{id}/outcome` |
    | **Book, confirm or cancel appointments** the agent made | `GET /appointments`, `PATCH /appointments/{id}` |
    | **Schedule callbacks** (single or in bulk, with idempotency keys) | `POST /callbacks`, `POST /callbacks/bulk`, `DELETE /callbacks?externalRef=` |
    | **Keep a do-not-call list** in sync with your compliance system | `GET /dnd`, `POST /dnd`, `POST /dnd/import`, `GET /dnd/check` |
    | **Receive real-time webhooks** for calls, appointments, callbacks and leads | `POST /webhooks`, `GET /webhooks/events` |

    ## Integrating from a CRM in 5 steps

    1. **Create an API key** — Settings → Developer → *Create API Key*. Give it only the scopes you need
...

...

...

...

...
        ]
...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /campaigns/{id}:
    get:
      summary: Get Campaign Details
      description: |
        Retrieve detailed information about a campaign including:
        - Current status and progress
        - Contact counts (total, called, remaining)
        - Schedule and timing information
        - Configuration settings
        
        <br />
        
        **Required Scope:** `campaigns:read`
      operationId: getCampaign
      tags: [Campaigns]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Campaign details with progress metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /webhooks:
    get:
      summary: List Webhooks
      description: |
        Retrieve all webhook subscriptions. Each webhook listens for specific events and delivers payloads to your endpoint.
        
        <br />
        
        **Webhook Status:**
        - `active` — Receiving events
        - `inactive` — Paused, not receiving events
        - `failed` — Too many delivery failures
        
        <br />
        
        **Required Scope:** `webhooks:read`
      operationId: listWebhooks
      tags: [Webhooks]
      responses:
        '200':
          description: List of webhook subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create Webhook
      description: |
        Subscribe to events with a new webhook. Your endpoint will receive HTTP POST requests when events occur.
        
        <br />
        
        **Setup Requirements:**
        - HTTPS endpoint (HTTP not allowed for security)
        - Endpoint must respond with 2xx within 30 seconds
        - Failed deliveries are retried with exponential backoff
        
        <br />
        
        **Signature Verification:**
        Each payload is signed with the `secret` returned here: `X-Zonvo-Signature` (legacy `X-AgentLabs-Signature`)
        holds the hex HMAC-SHA256 of the raw body. See the *Webhooks* section for a verification snippet.
        
        <br />
        
        **Required Scope:** `webhooks:write`
      operationId: createWebhook
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
            examples:
              callEvents:
                summary: Subscribe to call events
                value:
                  url: "https://your-server.com/webhooks/zonvo"
                  events: ["call.started", "call.completed", "call.failed"]
              crmSync:
                summary: CRM sync
                value:
                  name: "CRM sync"
                  url: "https://crm.example.com/hooks/zonvo"
                  events: ["call.completed", "lead.upserted", "appointment.booked", "appointment.cancelled", "callback.scheduled"]
                  description: "Zonvo → CRM"
      responses:
        '201':
          description: Webhook created with signing secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /webhooks/{id}:
    put:
      summary: Update Webhook
      description: |
        Modify a webhook's configuration. You can change the URL, events, or active status.
        
        <br />
        
        **Tip:** Set `active: false` to temporarily pause a webhook without deleting it.
        
        <br />
        
        **Required Scope:** `webhooks:write`
      operationId: updateWebhook
      tags: [Webhooks]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Webhook configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete Webhook
      description: |
        Permanently delete a webhook subscription. No further events will be delivered to this endpoint.
        
        **Required Scope:** `webhooks:write`
      operationId: deleteWebhook
      tags: [Webhooks]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Webhook deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /webhooks/{id}/test:
    post:
      summary: Test Webhook
      description: |
        Send a test payload to verify your webhook endpoint is configured correctly. The test payload mimics a real event.
        
        <br />
        
        **Test Process:**
        1. Zonvo AI sends a test payload to your URL
        2. Your server should return 2xx status
        3. Response includes success/failure and timing info
        
        <br />
        
        **Troubleshooting:**
        - Ensure your endpoint is publicly accessible
        - Check that you respond within 30 seconds
        - Verify SSL certificate is valid
        
        <br />
        
        **Required Scope:** `webhooks:write`
      operationId: testWebhook
      tags: [Webhooks]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Test results with timing and response info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookTestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token (API key)
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: object
        meta:
          $ref: '#/components/schemas/Meta'
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            deleted:
              type: boolean
              example: true
        meta:
          $ref: '#/components/schemas/Meta'
...
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [draft, scheduled, running, paused, completed, cancelled]
        agentId:
          type: string
          format: uuid
        phoneNumberId:
          type: string
          format: uuid
        totalContacts:
          type: integer
        calledContacts:
          type: integer
        scheduledStartTime:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CampaignResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Campaign'
        meta:
          $ref: '#/components/schemas/Meta'
...
    CreateWebhookRequest:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
          description: HTTPS endpoint to receive webhook payloads
        events:
          type: array
          minItems: 1
          items:
            type: string
            enum:
              - call.started
              - call.ringing
              - call.answered
              - call.completed
              - call.failed
              - call.transferred
              - call.no_answer
              - call.busy
              - call.voicemail
              - inbound_call.received
              - inbound_call.answered
              - inbound_call.completed
              - inbound_call.missed
              - campaign.started
              - campaign.paused
              - campaign.resumed
              - campaign.completed
              - campaign.failed
              - campaign.cancelled
              - appointment.booked
              - appointment.confirmed
              - appointment.cancelled
              - appointment.rescheduled
              - appointment.completed
              - appointment.no_show
              - callback.scheduled
              - lead.upserted
              - form.submitted
              - form.lead_created
              - flow.started
              - flow.completed
              - flow.failed
        secret:
          type: string
          minLength: 16
          maxLength: 64
          description: Custom signing secret, 16–64 characters (auto-generated if not provided)
        name:
          type: string
          maxLength: 255
        description:
          type: string
    UpdateWebhookRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        description:
          type: string
    Webhook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        url:
          type: string
        events:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    WebhookResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Webhook'
            - type: object
              properties:
                secret:
                  type: string
                  description: Only returned on creation
        meta:
          $ref: '#/components/schemas/Meta'
    WebhookListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Webhook'
        meta:
          $ref: '#/components/schemas/Meta'
...
    WebhookTestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            delivered:
              type: boolean
            statusCode:
              type: integer
            message:
              type: string
        meta:
          $ref: '#/components/schemas/Meta'

    # ---------------------------------------------------------------- Call outcomes
...